pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 120
target-version = ['py39']
exclude = '''
/(
  | build
  | .venv
  | out
)/
'''

[project]
name="opsmate"
long_description="Open source library to help automate the development and release cycle of API services and MLOps"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
authors = [
    { name = "databracket", email = "databracket9@gmail.com" },
]
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11"
]

dependencies = [
    "requests>=2.30.1,<3.0.0",
    "mlflow-skinny>=2.0.0,<3.0.0",

    # CLI libraries
    "click>=8.1.0,<9.0.0",
    "typer[all]==0.7.0",

    # Templates creation library
    "cookiecutter>2.1.0, <3.0.0",

    # File formats and model support
    "pyyaml>=6.0",
    "pydantic>=1.10.8,<2.0.0",
    "Jinja2>=2.11.2",

    # misc
    "cryptography>=41.0.0,<42.0.0",
    "aiohttp>=3.8.2",
]
dynamic = ["version"]

[tool.hatch.version]
path = "opsmate/__init__.py"

[tool.hatch.build.target.wheel.force-include]
"opsmate" = "opsmate"

[project.scripts]
opsmate = "opsmate.cli:entrypoint"

[project.urls]
Homepage = "https://github.com/Databracket9/OpsMate.git"
Repository = "https://github.com/Databracket9/OpsMate.git"
